move android implementation out
